+++ Snippets/Snippets_6-pre/packages.config
